Update test suite for expanded record types and runtime parity
Enhance RuntimeParityParserTests, RuntimeStructReaderTests, and
RuntimeNpcAutoDetectTests for new record types. Update NIF rendering
tests for skinning diagnostics and alpha composition. Update ESM parser
tests, script test helpers, and sample fixtures. Add xNVSE SDK
reference headers.

+++ tests/FalloutXbox360Utils.Tests/Core/Formats/Nif/Rendering/NifSkinningExtractorTests.cs
@@ -14,11 +14,11 @@ public sealed class NifSkinningExtractorTests
         {
             Bones =
             [
-                new NifBoneSkinInfo { VertexWeights = [((ushort)0, 0.50f)] },
-                new NifBoneSkinInfo { VertexWeights = [((ushort)0, 0.25f)] },
-                new NifBoneSkinInfo { VertexWeights = [((ushort)0, 0.15f)] },
-                new NifBoneSkinInfo { VertexWeights = [((ushort)0, 0.05f)] },
-                new NifBoneSkinInfo { VertexWeights = [((ushort)0, 0.05f)] }
+                new NifBoneSkinInfo { VertexWeights = [(0, 0.50f)] },
+                new NifBoneSkinInfo { VertexWeights = [(0, 0.25f)] },
+                new NifBoneSkinInfo { VertexWeights = [(0, 0.15f)] },
+                new NifBoneSkinInfo { VertexWeights = [(0, 0.05f)] },
+                new NifBoneSkinInfo { VertexWeights = [(0, 0.05f)] }
             ]
         };
